src/pyproject.toml
[project]
name = "open-data-lakehouse-demo"
version = "0.1.0"
description = "This webapp will walk through some setup and guidelines on building your own data lakehouse, based on open source technologies to enable multi-platform data lakehouse architecture"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "google-cloud-aiplatform>=1.94.0",
    "google-cloud-bigquery>=3.33.0",
    "google-cloud-core>=2.4.3",
    "google-cloud-managedkafka>=0.1.11",
    "google-cloud-storage>=2.19.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "sodapy>=2.2.0",
    "werkzeug>=3.1.3",
    "vega-datasets>=0.9.0",
    "ruff>=0.11.13",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
